create solar term selector class
